move server into app
