+++ server/mod.ts
@@ -1,0 +1,1 @@
+export type { AppType } from './impl/app.ts'
